% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_sentiment.R
\name{document_sentiment}
\alias{document_sentiment}
\title{Document Sentiment}
\usage{
document_sentiment(document)
}
\arguments{
\item{document}{character. The document that we want to analyze}
}
\value{
list
}
\description{
Analyze the sentiment of A document using exclusive sentiment lexicon
}
\examples{
\dontrun{
 document_sentiment ("Questa è una frase di esempio.")
}

}
