% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asin.R
\name{get_asin}
\alias{get_asin}
\title{Get ASIN}
\usage{
get_asin(url)
}
\arguments{
\item{url}{character. URL of the product on amazon.it}
}
\value{
character. The ASIN of the product.
}
\description{
Gives the ASIN of any product, by receiving its web address.
}
\examples{
\dontrun{
 get_asin("https://www.amazon.it/Tutti-gli-scritti-Platone/dp/8845290034")
}

}
