% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_review_pages.R
\name{number_of_review_pages}
\alias{number_of_review_pages}
\title{Number of the Review Pages}
\usage{
number_of_review_pages(number_of_reviews)
}
\arguments{
\item{number_of_reviews}{number. number of the reviews of the product on amazon.it}
}
\value{
numeric. The number of review pages for that product.
}
\description{
Gives the number of the review pages of any product, by receiving its number of reviews.
}
\examples{
\dontrun{
 number_of_review_pages(6329)
}

}
