% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_reviews.R
\name{number_of_reviews}
\alias{number_of_reviews}
\title{Number of the Reviews}
\usage{
number_of_reviews(asin)
}
\arguments{
\item{asin}{character. ASIN of the product.}
}
\value{
numerics. Number of the reviews written for the product on amazon.it
}
\description{
Gives the number of the reviews of any product, by receiving its ASIN.
}
\examples{
\dontrun{
 number_of_reviews("8845290034")
}

}
