% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_reviews.R
\name{scrape_reviews}
\alias{scrape_reviews}
\title{Scrape Reviews}
\usage{
scrape_reviews(asin, page_number)
}
\arguments{
\item{asin}{character. ASIN of the product.}

\item{page_number}{numeric. The page that user want to scrape reviews.}
}
\value{
data.frame
\describe{
  \item{title}{Title of the review}
  \item{text}{Content of the review}
  \item{star}{Number of stars}
}
}
\description{
Gives the title, content and given star of any product on amazon.it
}
\examples{
\dontrun{
 scrape_reviews("8845290034", 1)
}

}
