% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_analysis.R
\name{sentiment_analysis}
\alias{sentiment_analysis}
\title{Sentiment Analysis}
\usage{
sentiment_analysis(reviews)
}
\arguments{
\item{asin}{character. ASIN of the product.}

\item{page_number}{numeric. The page that user want to scrape reviews.}
}
\value{
data.frame
\describe{
  \item{title}{Title of the review}
  \item{text}{Content of the review}
  \item{star}{Number of stars}
}
}
\description{
Analyze the sentiment of reviews using exclusive sentiment lexicon and reports them
}
\examples{
\dontrun{
 sentiment_analysis("8845290034", 1)
}

}
