% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_sentiment.R
\name{term_sentiment}
\alias{term_sentiment}
\title{Term Sentiment}
\usage{
term_sentiment(word)
}
\arguments{
\item{document}{list The word that we are looking for its sentiment}
}
\value{
list
}
\description{
Analyze the sentiment of a word using exclusive sentiment lexicon
}
\examples{
\dontrun{
 term_sentiment ("perfetto")
}

}
